# api_interface_demo/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Sitewise Interface API
  description: Sitewise Interface API Draft001
  version: 1.0.2

  # TODO:
  #  - finish responses and validation
  #  - testing
servers:
# - url: http://Api114partial-env.eba-wam2eszv.eu-west-1.elasticbeanstalk.com
- url: /
paths:
...
  /zones:
    get:
      tags:
      - Zones
      operationId: get_zones
      responses:
        "200":
          description: "{'ResponseMetadata': {'RequestId': '0886febb-2290-4e17-8ddf-b208424d7399',\
            \ \n'HTTPStatusCode': 200, 'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 11:35:54\
            \ GMT', 'content-type': 'application/json',\n'content-length': '478',\
            \ 'connection': 'keep-alive', 'x-amzn-requestid': '0886febb-2290-4e17-8ddf-b208424d7399'},\
            \ \n'RetryAttempts': 0}, \n\n'assetSummaries': [\n{'id': '667ed4de-4cb4-42c3-a09b-77aad59ee4c3',\
            \ \n'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/667ed4de-4cb4-42c3-a09b-77aad59ee4c3',\
            \ \n'name': 'zone6', \n'assetModelId': '2bcfff28-ba76-44e5-8d8d-af10336fb448',\
            \ \n'creationDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),\
            \ \n'lastUpdateDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),\
            \ \n'status': {'state': 'ACTIVE'}, \n'hierarchies': [\n    {'id': 'b02d7852-f3e1-432f-9942-b9906451f82c',\
            \ 'name': 'SensorNodeChilds'}, \n    {'id': '9f909579-8fef-4772-bbbb-a880018775ac',\
            \ 'name': 'AssetSensorChilds'}]},\n    \n{'id': '667ed4de-4cb4-42c3-a09b-77aad59ee4c3',\
            \ \n'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/667ed4de-4cb4-42c3-a09b-77aad59ee4c3',\
            \ \n'name': 'zone7', \n'assetModelId': '2bcfff28-ba76-44e5-8d8d-af10336fb448',\
            \ \n'creationDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),\
            \ \n'lastUpdateDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),\
            \ \n'status': {'state': 'ACTIVE'}, \n'hierarchies': [\n    {'id': 'b02d7852-f3e1-432f-9942-b9906451f82c',\
            \ 'name': 'SensorNodeChilds'}, \n    {'id': '9f909579-8fef-4772-bbbb-a880018775ac',\
            \ 'name': 'AssetSensorChilds'}]}\n]\n}\n"
      x-openapi-router-controller: swagger_server.controllers.zones_controller
    post:
      tags:
      - Zones
      operationId: create_zone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameObj'
        required: true
      responses:
        "201":
          description: "{'ResponseMetadata': {'RequestId': 'a7f4441a-204a-490f-8abc-d8225d1a9abc',\
            \ \n'HTTPStatusCode': 202, \n'HTTPHeaders': {'date': 'Mon, 28 Dec 2020\
            \ 09:40:35 GMT', 'content-type': 'application/json', \n'content-length':\
            \ '184', 'connection': 'keep-alive', 'x-amzn-requestid': 'a7f4441a-204a-490f-8abc-d8225d1a9abc'},\
            \ \n'RetryAttempts': 0}, \n'assetId': '6627f48c-8d0a-481f-a305-b28e5258e319',\
            \ \n'assetArn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/6627f48c-8d0a-481f-a305-b28e5258e319',\
            \ \n'assetStatus': {'state': 'CREATING'}}\n"
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/{zoneId}:
    get:
      tags:
      - Zones
      operationId: get_zone_by_id
      parameters:
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: |
            {'ResponseMetadata': {'RequestId': 'a9606092-dbc1-4ef2-8db6-df6179dc25c8',
              'HTTPStatusCode': 200,
              'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 14:47:57 GMT',
               'content-type': 'application/json',
               'content-length': '1741',
               'connection': 'keep-alive',
               'x-amzn-requestid': 'a9606092-dbc1-4ef2-8db6-df6179dc25c8'},
              'RetryAttempts': 0},
             'assetId': '5d363986-a624-4e2b-a122-3ef01d6cde78',
             'assetArn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/5d363986-a624-4e2b-a122-3ef01d6cde78',
             'assetName': 'newzone333',
             'assetModelId': '2bcfff28-ba76-44e5-8d8d-af10336fb448',
             'assetProperties': [{'id': '6fc2fab2-1544-4455-bfa5-68ccb13e7be1',
               'name': 'ZoneName',
               'notification': {'topic': '$aws/sitewise/asset-models/2bcfff28-ba76-44e5-8d8d-af10336fb448/assets/5d363986-a624-4e2b-a122-3ef01d6cde78/properties/6fc2fab2-1544-4455-bfa5-68ccb13e7be1',
                'state': 'DISABLED'},
               'dataType': 'STRING'},
              {'id': 'a72b7466-02de-4295-8ac2-a858b3a4ad7c',
               'name': 'ZoneSensorNodeAvgTemp5min',
               'notification': {'topic': '$aws/sitewise/asset-models/2bcfff28-ba76-44e5-8d8d-af10336fb448/assets/5d363986-a624-4e2b-a122-3ef01d6cde78/properties/a72b7466-02de-4295-8ac2-a858b3a4ad7c',
                'state': 'DISABLED'},
               'dataType': 'DOUBLE'},
              {'id': 'fd61eac5-c93d-433c-9893-b037a4d1cbc3',
               'name': 'ZoneAssetSensorAvgTemp5min',
               'notification': {'topic': '$aws/sitewise/asset-models/2bcfff28-ba76-44e5-8d8d-af10336fb448/assets/5d363986-a624-4e2b-a122-3ef01d6cde78/properties/fd61eac5-c93d-433c-9893-b037a4d1cbc3',
                'state': 'DISABLED'},
               'dataType': 'DOUBLE'},
              {'id': '3fc34e60-ac4f-4395-aca7-2ddc36ebf284',
               'name': 'Zone Avg Power Usage (5min)',
               'notification': {'topic': '$aws/sitewise/asset-models/2bcfff28-ba76-44e5-8d8d-af10336fb448/assets/5d363986-a624-4e2b-a122-3ef01d6cde78/properties/3fc34e60-ac4f-4395-aca7-2ddc36ebf284',
                'state': 'DISABLED'},
               'dataType': 'DOUBLE'}],
             'assetHierarchies': [{'id': 'b02d7852-f3e1-432f-9942-b9906451f82c',
               'name': 'SensorNodeChilds'},
              {'id': '9f909579-8fef-4772-bbbb-a880018775ac', 'name': 'AssetSensorChilds'}],
             'assetCompositeModels': [],
             'assetCreationDate': datetime.datetime(2020, 12, 28, 12, 41, 28, tzinfo=tzlocal()),
             'assetLastUpdateDate': datetime.datetime(2020, 12, 28, 12, 41, 28, tzinfo=tzlocal()),
             'assetStatus': {'state': 'ACTIVE'}}
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
    delete:
      tags:
      - Zones
      operationId: delete_zone
      parameters:
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "204":
          description: "Entity removed successfully\n\n{'ResponseMetadata': {'RequestId':\
            \ '29725450-cf09-4a74-b510-a525c915834f', \n'HTTPStatusCode': 202, 'HTTPHeaders':\
            \ {'date': 'Mon, 28 Dec 2020 11:09:41 GMT', 'content-type': 'application/json',\
            \ \n'content-length': '36', 'connection': 'keep-alive', 'x-amzn-requestid':\
            \ '29725450-cf09-4a74-b510-a525c915834f'}, \n'RetryAttempts': 0}, \n'assetStatus':\
            \ {'state': 'DELETING'}}\n"
        "404":
          description: Entity with specified id was not found
        "406":
          description: Can not be delated because there are zones associated...
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/properties/:
    put:
      tags:
      - Zones
      operationId: put_zone_property_value
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetProperty'
        required: true
      responses:
        "200":
          description: |2

            {'ResponseMetadata': {'RequestId': 'e4209028-6785-4769-a639-ff51e6be8fb6',
            'HTTPStatusCode': 200,
            'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 10:56:17 GMT',
             'content-type': 'application/json',
             'content-length': '19',
             'connection': 'keep-alive',
             'x-amzn-requestid': 'e4209028-6785-4769-a639-ff51e6be8fb6'},
            'RetryAttempts': 0},
            'errorEntries': []}
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/{zoneId}/sensors:
    get:
      tags:
      - Zones
      operationId: get_zone_all_sensors
      parameters:
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: |
            Get zones sensors (asset sensors)

            {'ResponseMetadata': {'RequestId': '26e736bd-a7fc-4039-9699-e8a20838a4e0',
              'HTTPStatusCode': 200,
              'HTTPHeaders': {'date': 'Tue, 29 Dec 2020 11:14:37 GMT',
               'content-type': 'application/json',
               'content-length': '340',
               'connection': 'keep-alive',
               'x-amzn-requestid': '26e736bd-a7fc-4039-9699-e8a20838a4e0'},
              'RetryAttempts': 0},
             'assetSummaries': [
                {'id': '1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'name': 'sensornode6',
                   'assetModelId': '4c586961-6f99-4364-8847-267608f6c3e6',
                   'creationDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'lastUpdateDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'status': {'state': 'ACTIVE'},
                   'hierarchies': []},
                {'id': '1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'name': 'assetsensor7',
                   'assetModelId': '4c586961-6f99-4364-8847-267608f6c3e6',
                   'creationDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'lastUpdateDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'status': {'state': 'ACTIVE'},
                   'hierarchies': []}
               ]
             }
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/{zoneId}/sensornodesensors:
    get:
      tags:
      - Zones
      operationId: get_zone_sensornode_sensors
      parameters:
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: |
            Get zones sensors (sensor node)

            {'ResponseMetadata': {'RequestId': '26e736bd-a7fc-4039-9699-e8a20838a4e0',
              'HTTPStatusCode': 200,
              'HTTPHeaders': {'date': 'Tue, 29 Dec 2020 11:14:37 GMT',
               'content-type': 'application/json',
               'content-length': '340',
               'connection': 'keep-alive',
               'x-amzn-requestid': '26e736bd-a7fc-4039-9699-e8a20838a4e0'},
              'RetryAttempts': 0},
             'assetSummaries': [
                {'id': '1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'name': 'sensornode6',
                   'assetModelId': '4c586961-6f99-4364-8847-267608f6c3e6',
                   'creationDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'lastUpdateDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'status': {'state': 'ACTIVE'},
                   'hierarchies': []},
                {'id': '1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'name': 'sensornode7',
                   'assetModelId': '4c586961-6f99-4364-8847-267608f6c3e6',
                   'creationDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'lastUpdateDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'status': {'state': 'ACTIVE'},
                   'hierarchies': []}
               ]
             }
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/{zoneId}/assetsensorsensors:
    get:
      tags:
      - Zones
      operationId: get_zone_assetsensor_sensors
      parameters:
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: |
            Get zones sensors (asset sensors)

            {'ResponseMetadata': {'RequestId': '26e736bd-a7fc-4039-9699-e8a20838a4e0',
              'HTTPStatusCode': 200,
              'HTTPHeaders': {'date': 'Tue, 29 Dec 2020 11:14:37 GMT',
               'content-type': 'application/json',
               'content-length': '340',
               'connection': 'keep-alive',
               'x-amzn-requestid': '26e736bd-a7fc-4039-9699-e8a20838a4e0'},
              'RetryAttempts': 0},
             'assetSummaries': [
                {'id': '1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'name': 'assetsensor6',
                   'assetModelId': '4c586961-6f99-4364-8847-267608f6c3e6',
                   'creationDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'lastUpdateDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'status': {'state': 'ACTIVE'},
                   'hierarchies': []},
                {'id': '1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'arn': 'arn:aws:iotsitewise:eu-west-1:241091466505:asset/1f08d927-c544-48f5-a9d5-c98bd9ce852f',
                   'name': 'assetsensor7',
                   'assetModelId': '4c586961-6f99-4364-8847-267608f6c3e6',
                   'creationDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'lastUpdateDate': datetime.datetime(2020, 12, 23, 11, 5, 59, tzinfo=tzlocal()),
                   'status': {'state': 'ACTIVE'},
                   'hierarchies': []}
               ]
             }
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/properties/{zoneId}:
    get:
      tags:
      - Zones
      operationId: get_zone_property_values
      parameters:
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: "Get zone property values\n\n{'ZoneName': {'value': {'stringValue':\
            \ 'Zone 0'}, 'timestamp': {'timeInSeconds': 1607011700, 'offsetInNanos':\
            \ 0}, 'quality': 'GOOD'}, \n'ZoneSensorNodeAvgTemp5min': {'value': {'doubleValue':\
            \ '23.0'}, 'timestamp': {'timeInSeconds': 1607011700, 'offsetInNanos':\
            \ 0}, 'quality': 'GOOD'}, \n'ZoneAssetSensorAvgTemp5min': {'value': {'doubleValue':\
            \ '23.0'}, 'timestamp': {'timeInSeconds': 1607011700, 'offsetInNanos':\
            \ 0}, 'quality': 'GOOD'}, \n'Zone Avg Power Usage (5min)': {'value': {'doubleValue':\
            \ '23.0'}, 'timestamp': {'timeInSeconds': 1607011700, 'offsetInNanos':\
            \ 0}, 'quality': 'GOOD'}}\n\n# NOTE:\n#  - if the property hasn't received\
            \ any value yet from the sensor\n#    the response will be 'null' like\
            \ this:\n  'Zone Avg Power Usage (5min)': 'null'\n  \n"
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/properties/history:
    post:
      tags:
      - Zones
      operationId: get_zone_property_value_history
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetPropertyValueHistoryRequest'
        required: true
      responses:
        "200":
          description: |
            Get property values from a specified time range.
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/{zoneId}/associate/{sensorId}:
    post:
      tags:
      - Zones
      operationId: associate_sensor
      parameters:
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      - name: sensorId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: |
            Associate a senor to a zone

            {'ResponseMetadata': {'RequestId': 'f0d2f325-59fb-4b14-93e9-5a565383eae2',
            'HTTPStatusCode': 200,
            'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 14:23:27 GMT',
             'content-type': 'application/json',
             'content-length': '0',
             'connection': 'keep-alive',
             'x-amzn-requestid': 'f0d2f325-59fb-4b14-93e9-5a565383eae2'},
            'RetryAttempts': 0}}
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
  /zones/{zoneId}/disassociate/{sensorId}:
    post:
      tags:
      - Zones
      operationId: disassociate_sensor
      parameters:
      - name: zoneId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      - name: sensorId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        "200":
          description: |
            Disassociate a sensor from a zone

            {'ResponseMetadata': {'RequestId': 'f0d2f325-59fb-4b14-93e9-5a565383eae2',
            'HTTPStatusCode': 200,
            'HTTPHeaders': {'date': 'Mon, 28 Dec 2020 14:23:27 GMT',
             'content-type': 'application/json',
             'content-length': '0',
             'connection': 'keep-alive',
             'x-amzn-requestid': 'f0d2f325-59fb-4b14-93e9-5a565383eae2'},
            'RetryAttempts': 0}}
        "404":
          description: Entity with specified id was not found
      x-openapi-router-controller: swagger_server.controllers.zones_controller
...
components:
  schemas:
...
    AssetProperty:
      type: object
      properties:
        propertyName:
          type: string
        assetId:
          type: string
        dataType:
          type: string
          enum: [ 'integerValue', 'stringValue', 'doubleValue', 'booleanValue' ]
        value:
          # TODO
          type: string
    Id:
      type: string
...
    NameObj:
      type: object
      properties:
        Name:
          type: string
...
    AssetPropertyValueHistoryRequest:
      type: object
      required:
      - assetId
      - startDate
      - endDate
      properties:
        assetId:
          type: string
        maxResults:
          type: number
          example: 50
        startDate:
          type: string
          format: datetime
          pattern: ([0-9]{4})-([0-9]{2})-([0-9]{2})
          example: "2020-12-20"
        endDate:
          type: string
          format: datetime
          pattern: ([0-9]{4})-([0-9]{2})-([0-9]{2})
          example: "2020-12-30"
        timeOrdering:
          type: string
          enum: ["ASCENDING", "DESCENDING"]
        include_prop:
          type: array
          items:
            type: string
          example: ["Temperature", "WorkingState"]


####################### SECURITY
  securitySchemes:
    basicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
      x-basicInfoFunc: application.basic_auth
security:
  - basicAuth: ['secret']  # <-- use the same name here
#  - basicAuth: ['secret']  # <-- use the same name here

#   securitySchemes:
#     oAuth2Password:
#       type: oauth2
#       description: See https://developers.getbase.com/docs/rest/articles/oauth2/requests
#       flows:
#         password:
#           tokenUrl: https://api.getbase.com/oauth2/token
#           scopes: {}
# security:
#   - oAuth2Password: []


#  responses:
#    TODO!!!